Back refatoracao

+++ src/main/java/com/consol/api/service/DespesaService.java
@@ -1,17 +1,12 @@
 package com.consol.api.service;
 
-import com.consol.api.dto.despesa.DespesaAtualizarDto;
-import com.consol.api.dto.despesa.DespesaCadastroDto;
-import com.consol.api.dto.despesa.DespesaConsultaDto;
-import com.consol.api.dto.despesa.DespesaMapper;
 import com.consol.api.entity.Despesa;
+import com.consol.api.entity.Familia;
+import com.consol.api.entity.exception.EntidadeNaoEncontradaException;
+import com.consol.api.entity.exception.RequisicaoIncorretaException;
 import com.consol.api.repository.DespesaRepository;
-import com.consol.api.repository.DonatarioRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -21,26 +16,22 @@ import java.util.Optional;
 public class DespesaService {
 
     private final DespesaRepository despesaRepository;
-
     private final FamiliaService familiaService;
 
-    public List<Despesa> listar() {
-        return despesaRepository.findAll();
+    public Despesa salvar(Despesa despesa, int idFamilia) {
+        Familia familia = familiaService.porId(idFamilia);
+
+        if (familia == null) throw new RequisicaoIncorretaException("Família");
+
+        despesa.setFamilia(familia);
+        Despesa despesaSalva = despesaRepository.save(despesa);
+        return despesaSalva;
     }
 
     public Despesa buscarPorId(Integer id) {
-
         return despesaRepository.findById(id).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.NOT_FOUND)
+                () -> new EntidadeNaoEncontradaException("Despesa")
         );
-    }
-
-    public Despesa salvar(Despesa despesa) {
-        if (despesa == null) return null;
-
-        Despesa despesaSalva = despesaRepository.save(despesa);
-
-        return despesaSalva;
     }
 
     public List<Despesa> listarPorFamilia(int idFamilia) {
@@ -48,22 +39,30 @@ public class DespesaService {
         return despesaRepository.findByFamiliaId(idFamilia);
     }
 
-    public Despesa atualizarDespesa(Integer id, Despesa despesa){
-        Optional<Despesa> despesaBuscadaOpt = despesaRepository.findById(id);
-        if (despesaBuscadaOpt.isEmpty()) return null;
+    public Despesa atualizarDespesa(Despesa despesa, Integer id){
+        Optional <Despesa> despesaBanco = despesaRepository.findById(id);
 
-        Despesa despesaBuscada = despesaBuscadaOpt.get();
-        despesa.setId(id);
+        if (despesaBanco == null) throw new EntidadeNaoEncontradaException("Despesa");
 
-        if (despesa.getTipo() == null) despesa.setTipo(despesaBuscada.getTipo());
-        if (despesa.getGasto() == null) despesa.setGasto(despesaBuscada.getGasto());
+        Despesa despesaAtualizar = despesaBanco.get();
+
+        if (despesa.getTipo() != null && !despesa.getTipo().equals("") && !despesa.getTipo().equals(" ")) despesaAtualizar.setTipo(despesa.getTipo());
+        if (despesa.getGasto() != null) despesaAtualizar.setGasto(despesa.getGasto());
 
 
-        return despesaRepository.save(despesa);
+        return despesaRepository.save(despesaAtualizar);
+
     }
 
     public void deletarPorId(Integer id) {
-        if (!despesaRepository.existsById(id));
+        Boolean exist = despesaRepository.existsById(id);
+
+        if (!exist) throw new EntidadeNaoEncontradaException("Despesa");
         despesaRepository.deleteById(id);
+
+    }
+
+    public List<Despesa> listar(){
+        return despesaRepository.findAll();
     }
 }
